requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
branca==0.6.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
django-location-field==2.1.0
django-ranged-response==0.2.0
folium==0.13.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
neo4j==5.2.0
numpy==1.23.4
Pillow==9.3.0
pytz==2022.6
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
